Updating requirements for install
At least presently on my 10.15.7 MBP, I get an error 

```
wallpaper: A full installation of Xcode.app 13.0 is required to compile
this software. Installing just the Command Line Tools is not sufficient.

Xcode 13.0 cannot be installed on macOS 10.15.
You must upgrade your version of macOS.
```

so I upgrade the requirements for others in the future.